refactor: extract card mapping to shared module
Consolidate getRarityLabel, getTagLabel, getShortName, getShortCity,
and getBonusText from render.ts and certificate.ts into
src/shared/card-mapper.ts.

- Removes ~108 lines of duplicate inline logic from render.ts
- Removes ~13 lines from certificate.ts' standalone getTagLabel
- Card mapper centralizes all manual overrides, labels, and bonus text
  generation in one testable module
